+++ src/main/webui/src/app/app.config.ts
@@ -1,0 +1,23 @@
+/*
+  Copyright 2022-2026 VALAWAI
+
+  Use of this source code is governed by GNU General Public License version 3
+  license that can be found in the LICENSE file or at
+  https://opensource.org/license/gpl-3-0/
+*/
+
+import { ApplicationConfig, provideZoneChangeDetection } from '@angular/core';
+import { provideRouter } from '@angular/router';
+
+import { APP_ROUTES } from './app.routes';
+import { provideAnimationsAsync } from '@angular/platform-browser/animations/async';
+import { provideHttpClient } from '@angular/common/http';
+
+export const appConfig: ApplicationConfig = {
+	providers: [
+		provideZoneChangeDetection({ eventCoalescing: true }), 
+		provideRouter(APP_ROUTES), 
+		provideAnimationsAsync(),
+		provideHttpClient()
+	]
+};