+++ src/main/webui/src/app/main/main.module.ts
@@ -9,7 +9,6 @@
 import { NgModule } from '@angular/core';
 import { CommonModule } from '@angular/common';
 
-import { MainRoutingModule } from './main-routing.module';
 import { MainComponent } from './main.component';
 import { MatIconModule } from '@angular/material/icon';
 import { MatButtonModule } from '@angular/material/button';
@@ -22,7 +21,6 @@ import { MatMenuModule } from '@angular/material/menu';
 	],
 	imports: [
 		CommonModule,
-		MainRoutingModule,
 		MatIconModule,
 		MatButtonModule,
 		MatMenuModule